Added texture Offset
Textures can now be offset by a certain amount of pixels so that they can touch in the center. This is controlled currently with a slider that gives the same offset for all items, but will be changeable later with regions on each item so they can each have different offsets, and eventually, offsets will ideally be set automatically with a getPixelColor method.