Items Now Have Individual offsets
Offsets are now controlled by a region input and no longer affect other items on the same page. However, there are now a few new issues because of this: More items added now cause increasingly more lag, and once a second page is reached, the items have the same offset as the item at the same position on the other pages, and the regions on the other page disappear.